api/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Email & SMS
aiosmtplib==3.0.1
jinja2==3.1.2

# Storage
boto3==1.29.7
minio==7.2.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Development
black==23.11.0
ruff==0.1.6
